requirements.txt
flwr==1.19.0
matplotlib==3.10.3
netron==8.1.9
numpy==1.23.5
pandas==2.3.0
Pillow==11.3.0
scikit_learn==1.7.0
seaborn==0.13.2
segmentation_models==1.0.1
shap==0.48.0
tensorflow==2.12.0
torch==2.6.0
torchvision==0.21.0
torchviz==0.0.3
xgboost==3.0.2
